Create new search function for summary panel enhancements changes
Creating a new wrapper function in the data explorer client instance for searching that utilizes the backend for searching and sorting. The backend powered search is being implemented in an isolated way to avoid regressions in the data grid column filtering functionality which currently works.

As part of adding search and sort, the order the data
should be rendered in (with regards to pinned columns as well)
needs to be maintained separately from the cache.

The cache previously only contained the data we needed to render and was used as the source of truth on what to render
and the order to render it.

Now we utilze an array to determine what to render and in what order